corrected a lot of compiler error messages and warnings, added preconditioning of updated SVD

+++ src/cplscheme/config/CouplingSchemeConfiguration.cpp
@@ -1066,7 +1066,7 @@ bool CouplingSchemeConfiguration:: checkIfDataIsCoarse
   isCoarse = not utils::contained(id, fineIDs);
 
   std::cout<<" fineIDs = [";
-  for(int i = 0; i < fineIDs.size();i++){
+  for(int i = 0; i < (int)fineIDs.size();i++){
     std::cout<<fineIDs.at(i)<<", ";
   }
   std::cout<<"], id: "<<id<<std::endl;
