+++ src/cplscheme/config/PostProcessingConfiguration.cpp
@@ -332,7 +332,7 @@ void PostProcessingConfiguration:: xmlEndTagCallback
       else{
         // no preconditioner defined
         std::vector<double> factors;
-        for (int id = 0; id < _config.dataIDs.size(); ++id) {
+        for (int id = 0; id < (int)_config.dataIDs.size(); ++id) {
           factors.push_back(1.0);
         }
         _preconditioner = impl::PtrPreconditioner (new impl::ConstantPreconditioner(dims, factors));
